requirements.txt
# Dependências principais do AgentOS
fastapi>=0.109.2,<0.110.0
uvicorn[standard]>=0.22.0,<0.25.0
pydantic>=2.7.3,<3.0.0  # Atualizado para compatibilidade com mem0ai
python-dotenv>=1.0.1
requests>=2.32.3,<3.0.0

# Modelos de IA
openai>=1.33.0,<2.0.0

# Memória persistente e RAG
mem0ai==0.1.0
pinecone-client==3.0.0
supabase>=2.0.0,<3.0.0

# Utilitários
numpy>=1.26.2,<2.0.0
pandas>=2.0.3,<3.0.0
loguru==0.7.0

# Dependências de desenvolvimento
pytest==7.4.0
pytest-asyncio==0.21.0
black==23.1.0
flake8==6.0.0

# Produção
gunicorn>=20.1.0,<21.0.0

# Nota: agno[pinecone] removido devido a problemas de compatibilidade
# Instalar manualmente se necessário: pip install agno[pinecone]